* Fixed compilation warning with data converting
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@1697 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ combat.cpp
@@ -84,16 +84,16 @@ bool Combat::getMinMaxValues(Creature* creature, Creature* target, int32_t& min,
 				{
 					if(vocation && vocation->getMagicHealingMultiplier() != 1.0)
 					{
-						min *= vocation->getMagicHealingMultiplier();
-						max *= vocation->getMagicHealingMultiplier();
+						min *= (int32_t)vocation->getMagicHealingMultiplier();
+						max *= (int32_t)vocation->getMagicHealingMultiplier();
 					}
 				}
 				else //Attack multipler
 				{
 					if(vocation && vocation->getMagicMultiplier() != 1.0)
 					{
-						min *= vocation->getMagicMultiplier();
-						max *= vocation->getMagicMultiplier();
+						min *= (int32_t)vocation->getMagicMultiplier();
+						max *= (int32_t)vocation->getMagicMultiplier();
 					}
 				}
 				return true;
